refactor(suite): move fee modules to @suite-common/fee and @suite/fee (incremental)
