+++ packages/suite/package.json
@@ -38,6 +38,7 @@
         "@suite-common/earn-stablecoin": "workspace:^",
         "@suite-common/earn-stablecoin-api": "workspace:*",
         "@suite-common/earn-staking-api": "workspace:*",
+        "@suite-common/fee": "workspace:*",
         "@suite-common/feedback": "workspace:*",
         "@suite-common/fiat-services": "workspace:*",
         "@suite-common/firmware": "workspace:*",
@@ -91,6 +92,7 @@
         "@suite/experimental": "workspace:*",
         "@suite/external-links": "workspace:*",
         "@suite/feature-feedback": "workspace:*",
+        "@suite/fee": "workspace:*",
         "@suite/firmware-upgrade": "workspace:*",
         "@suite/flags": "workspace:*",
         "@suite/github": "workspace:*",
