+++ packages/suite/src/components/wallet/Fees/CollapsibleFees/CustomFee/CustomFeeTron.tsx
@@ -2,6 +2,7 @@ import { useEffect } from 'react';
 import { useFormContext, useFormState } from 'react-hook-form';
 import { useSelector } from 'react-redux';
 
+import { FEE_LIMIT } from '@suite/fee';
 import { Translation, useTranslation } from '@suite/intl';
 import { selectLanguage } from '@suite/settings';
 import { type FormState } from '@suite-common/wallet-types';
@@ -10,8 +11,6 @@ import { NumberInput } from '@trezor/product-components';
 import { BigNumber } from '@trezor/utils';
 
 import { InputError } from 'src/components/wallet';
-
-import { FEE_LIMIT } from './constants';
 
 export const CustomFeeTron = () => {
     const locale = useSelector(selectLanguage);
