+++ packages/suite/src/components/wallet/Fees/CollapsibleFees/MaximumFee.tsx
@@ -1,13 +1,11 @@
 import { Translation } from '@suite/intl';
+import { type TransactionMaxFee, useFeesContext } from '@suite-common/fee';
 import { selectAreFeesLoading } from '@suite-common/wallet-core';
 import { Column, LoadingContent, Text } from '@trezor/components';
 import { type TypographyStyle } from '@trezor/theme';
 
 import { BaseCurrencyValue, FormattedCryptoAmount } from 'src/components/suite';
 import { useSelector } from 'src/hooks/suite';
-
-import { useFeesContext } from '../context/FeesContext';
-import { type TransactionMaxFee } from './hooks/useTransactionMaxFee';
 
 export type MaximumFeeProps = {
     typographyStyle: TypographyStyle;
