+++ packages/suite/src/components/wallet/Fees/DustPreventionNotice.tsx
@@ -1,14 +1,13 @@
 import { useWatch } from 'react-hook-form';
 
 import { Translation } from '@suite/intl';
+import { useFeesContext } from '@suite-common/fee';
 import { selectAreFeesLoading } from '@suite-common/wallet-core';
 import { type FormState } from '@suite-common/wallet-types';
 import { Note } from '@trezor/components';
 import { BigNumber } from '@trezor/utils';
 
 import { useSelector } from 'src/hooks/suite';
-
-import { useFeesContext } from './context/FeesContext';
 
 type DustPreventionNoticeProps = {
     chosenFeePerByte: string | undefined;
