+++ packages/suite/src/components/wallet/Fees/index.tsx
@@ -2,22 +2,20 @@ import { CurrentFee } from './CollapsibleFees/CustomFee/CurrentFee';
 import { CustomFee } from './CollapsibleFees/CustomFee/CustomFee';
 import { CustomFeeEthereum } from './CollapsibleFees/CustomFee/CustomFeeEthereum';
 import { CustomFeeMisc } from './CollapsibleFees/CustomFee/CustomFeeMisc';
-import { CustomFeeTooLowBanner } from './CollapsibleFees/CustomFee/CustomFeeTooLowBanner';
 import { BitcoinFeeCards } from './CollapsibleFees/StandardFee/BitcoinFeeCards';
 import { EthereumFeeCards } from './CollapsibleFees/StandardFee/EthereumFeeCards';
-import { FeeCard } from './CollapsibleFees/StandardFee/FeeCard';
 import { MiscFeeCards } from './CollapsibleFees/StandardFee/MiscFeeCards';
 import { StandardFee } from './CollapsibleFees/StandardFee/StandardFee';
+
+export { CustomFeeTooLowBanner, FeeCard } from '@suite/fee';
 
 export {
     CustomFee,
     CurrentFee,
     CustomFeeEthereum,
     CustomFeeMisc,
-    CustomFeeTooLowBanner,
     BitcoinFeeCards,
     EthereumFeeCards,
-    FeeCard,
     MiscFeeCards,
     StandardFee,
 };
