+++ suite-common/fee/src/useTransactionMaxFee.ts
@@ -8,7 +8,7 @@ import {
 } from '@suite-common/wallet-utils';
 import { BigNumber } from '@trezor/utils';
 
-import { type FeesContextType } from '../../context/FeesContext';
+import { type FeesContextType } from './FeesContext';
 
 export type TransactionMaxFeeProps = Pick<FeesContextType, 'networkSymbol' | 'composedLevels'> & {
     selectedFeeLevel?: FeesContextType['selectedFeeLevel'];
