+++ suite-common/fee/tsconfig.json
@@ -1,5 +1,11 @@
 {
     "extends": "../../tsconfig.base.json",
     "compilerOptions": { "outDir": "libDev" },
-    "references": []
+    "references": [
+        { "path": "../wallet-config" },
+        { "path": "../wallet-types" },
+        { "path": "../wallet-utils" },
+        { "path": "../../packages/connect" },
+        { "path": "../../packages/utils" }
+    ]
 }
