+++ suite/fee/src/FeeCardsWrapper.ts
@@ -4,9 +4,6 @@ import { spacingsPx } from '@trezor/theme';
 
 import { FEE_CARD_MIN_WIDTH } from './FeeCard';
 
-// Can't be in `StandardFee` component because of circular dependency:
-// - `StandardFee` imports `BitcoinFeeCards`, `EthereumFeeCards`, `MiscFeeCards`
-// - `BitcoinFeeCards`, `EthereumFeeCards`, `MiscFeeCards` import `FeeCardsWrapper`
 export const FeeCardsWrapper = styled.div`
     width: 100%;
     display: grid;
