+++ suite/fee/src/customFee/CustomFeeBasicProps.ts
