+++ suite/fee/src/customFee/CustomFeeTooLowBanner.tsx
@@ -3,14 +3,14 @@ import { useWatch } from 'react-hook-form';
 
 import { LearnMoreButton } from '@suite/external-links';
 import { Translation } from '@suite/intl';
+import { useFeesContext } from '@suite-common/fee';
 import { type FormState } from '@suite-common/wallet-types';
 import { getLowestFeeFromLevels } from '@suite-common/wallet-utils';
 import { Banner, Collapsible } from '@trezor/components';
 import { HELP_CENTER_TRANSACTION_FEES_URL } from '@trezor/urls';
 import { BigNumber } from '@trezor/utils';
 
-import { FEE_PER_UNIT } from './constants';
-import { useFeesContext } from '../../context/FeesContext';
+import { FEE_PER_UNIT } from './customFeeConstants';
 
 export const CustomFeeTooLowBanner = memo(function CustomFeeTooLowBannerInner() {
     const { feeInfo } = useFeesContext();
