+++ suite/fee/src/feeLevelTranslationMap.ts
